Split SQLite storage into schema and serde mixins
Schema creation, version migration and backfill were 552 of sqlite.py's 1647
lines. Move them out with the schema constants, and move row serialisation into
its own mixin. The push history dataclasses move to records.py so the serde
mixin does not import sqlite.py back; sqlite.py re-exports them and the schema
constants so the old paths still resolve.

Pure move: SQLiteStorage keeps all 55 members with identical bodies. The
_locked_sqlite_method wrapper list still resolves through the MRO.

Also drop the orphaned delivered_targets serialisers; no such column exists.
